fix multiple definitions.

+++ python/jittor/src/ctorch/c10/util/Half.h
@@ -6,10 +6,11 @@
 #include "ops/op_register.h"
 #include "pyjt/py_converter.h"
 #include "misc/cuda_flags.h"
+#include "type/fp16_compute.h"
 #include <cuda_runtime.h>
 #include "helper_cuda.h"
 
 
 namespace c10 {
-    using Half = jittor::float32; // to do: change this to float16.
+    using Half = jittor::float16; 
 }
